requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
ase==3.20.1
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
bleach==3.2.1
bump2version==0.5.11
certifi==2020.12.5
chardet==4.0.0
coverage==4.5.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
future==0.18.2
gemmi==0.4.4
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==5.7.0
Jinja2==2.11.2
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mendeleev==0.2.17
monty==2021.3.3
more-itertools==8.6.0
mpmath==1.2.1
networkx==2.5.1
numpy==1.19.2
packaging==20.8
palettable==3.3.0
pandas==1.1.2
pathlib==1.0.1
pathtools==0.1.2
persist-queue==0.6.0
Pillow==8.2.0
pkginfo==1.6.1
plotly==4.14.3
pluggy==0.13.1
ppinterpf @ file:///home/benjamin/Code/ppinterpf
psutil==5.8.0
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.1.1
Pygments==2.7.3
pylint==2.6.0
pymatgen==2020.12.3
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
PyYAML==5.3.1
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.6.2
semantic-version==2.8.5
setuptools-rust==0.11.5
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
spglib==1.16.1
Sphinx==1.8.5
sphinx-rtd-theme==1.0.0
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
SQLAlchemy==1.3.19
sympy==1.7.1
tabulate==0.8.9
threadpoolctl==2.1.0
toml==0.10.2
tox==3.14.0
tqdm==4.55.0
twine==1.14.0
uncertainties==3.1.5
unitcellsampling @ file:///home/benjamin/Code/unitcellsampling
urllib3==1.26.4
virtualenv==20.2.2
watchdog==0.9.0
webencodings==0.5.1
wrapt==1.12.1
